Fix timbre program bank decoding to match Kronos PcgId mapping
Based on C# KronosProgramBanks.cs, the bank byte (PcgId) mapping is:
- I-A through I-F: 0-5
- GM: 6
- U-A through U-G: 17-23
- U-AA through U-GG: 24-30

Previously we were using a simple 0-7 for internal and 0x20+ for user,
which was incorrect. Now timbres correctly show U-F, U-G, etc.

Note: The referenced programs may not be found if those banks aren't
parsed yet, but the bank reference is now correct.